requirements.txt
flask
gunicorn
ultralytics
Pillow
requests
torch
torchvision
